twitter-upload/requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
requests-oauthlib==1.3.1
oauthlib==3.2.2
